requirements.txt
streamlit
opencv-python
ultralytics
pillow
numpy
